hw3/requirements.txt
numpy==1.19.2
torch==1.6.0
torchvision==0.7.0
opencv-python==4.4.0.44
matplotlib==3.3.2
Pillow==7.2.0
yacs==0.1.8